landman/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.11
# To update, run:
#
#    pip-compile
#
asgiref==3.6
    # via django
django==4.1.3
    # via -r requirements.in
django-admin-list-filter-dropdown
    # https://django-baton.readthedocs.io/en/stable/list_filters.html
django-simple-history
    # https://django-simple-history.readthedocs.io/en/latest/
django-smart-selects
    # https://django-smart-selects.readthedocs.io/en/latest/
faker==15.3.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via faker
six==1.16.0
    # via python-dateutil
sqlparse==0.4.3
    # via django
